requirements.txt
streamlit==0.63.0
Pillow==7.2.0
